Slight refactor for clarity.
