+++ src/main/java/nl/tudelft/lifetiles/graph/controller/TileController.java
@@ -3,9 +3,9 @@ package nl.tudelft.lifetiles.graph.controller;
 import java.util.ArrayList;
 import java.util.List;
 
-import nl.tudelft.lifetiles.graph.models.Tile;
+import nl.tudelft.lifetiles.graph.view.Tile;
 import nl.tudelft.lifetiles.graph.view.TileView;
-import nl.tudelft.lifetiles.graph.view.Vertex;
+import nl.tudelft.lifetiles.graph.view.VertexView;
 import javafx.scene.Group;
 import javafx.scene.paint.Color;
 
@@ -45,7 +45,7 @@ public class TileController {
      * @param v
      *            - to be changed vertex
      */
-    public final void changeColour(final Color color, final Vertex v) {
+    public final void changeColour(final Color color, final VertexView v) {
         List<Object> args = new ArrayList<Object>();
         args.add(v);
         args.add(color);
@@ -73,7 +73,7 @@ public class TileController {
             final List<Object> args) {
 
         if (propertyName.equals("COLOR_CHANGED")) {
-            viewVar.changeVertexColour((Vertex) args.get(0),
+            viewVar.changeVertexColor((VertexView) args.get(0),
                     (Color) args.get(1));
         }
     }
